Add cluster.raft.spi.EntryData et al.
We need to reign in ReplicatedLogEntry serialization, so that we can
operate without Serializable being in the picture.

Baseline semantincs of a ReplicateLogEntry is that it represents a step
in RAFT maintentenance, affecting either the contents of StateSnapshot
or a RAFT server state transition.

Introduce cluster.raft.spi.EntryData as the unifying concept, with
RaftDelta representing RAFT server transitions and StateDelta
representing user state transitions.

JIRA: CONTROLLER-2137
Change-Id: Icbf3b32a9ca68c6b60709ea79b509b3078793ea9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>