Ditch use of ByteSource
Rename InputStreamProvider to StreamSource with two specializations, based
on whether or not we know the size.

Allow conversion from StreamSource to SizedStreamSource, for the cost of
consuming the input stream once.

Also disconnect SnapshotSource from StreamSource: the former indicates
the format of the underlying InputStreams, while StreamSource provides
the streams themselves.

JIRA: CONTROLLER-2134
Change-Id: I214b990c9b9d573661de824cbb5d371bcd66ccb7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>