Refactor FileBackedOutputStream
The lifecycle here is a bit hairy. Improve it by keeping track of what
state we are in.

While we are in the area, split out TransientFile(StreamSource), which
is backed by a temporary file.

Major improvement here is TransientFile, which acts as the GC anchor to
a file -- eliminating the need to hold on to FileBackedOutputStream.

JIRA: CONTROLLER-2134
Change-Id: I3fbdcca3e927551572a38dd4d0268b0d56b68aa9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>