Add EntryStore and LogEntry
Split out the methods supporting entries and define a baseline for how a
LogEntry looks like.

Naming is adjusted to match the RAFT paper, i.e. a LogEntry carries a
'command', not 'data'. This paves the way for a potential EntrySource
interface.

This clarifies RaftActorSnapshotCohort's relationship with
StateSnapshot, introducing a StateSnapshot.Support to hold the
indirection to reader/writer.

JIRA: CONTROLLER-2137
Change-Id: I1513ab5ba8533bc5816294b15a4c5fcc19b974a2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>