Add ReplicatedLog.lookupStoredMeta()
AbstractLeader does not need the full ReplicatedLogEntry, it just needs
its metadata with a durability flag.

Eliminate a caller for ReplicateLog.get() with a dedicated method
needing only EntryMeta.

Since we are in the area, turn a TODO in consensus reaching into a FIXME
we a better analysis.

JIRA: CONTROLLER-2137
Change-Id: Ifa593ddfefcf3015502e8b42e30c37afb85768de
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>