Split out DeleteEntries handling
ReplicateLog is using persist(DeleteEntries) to talk directly to Pekko
peristence to undo effects of previous persist(ReplicatedLogEntry)
calls.

Document this use and introduce an explicit
ReplicatedLog.trimToReceive() method instead of the confusing
removeFrom() and removeFromAndPersist() methods.

removeFrom() is called from tests and recovery only. It is renamed to
removeRecoveredEntries() called only from recovery. The implementation
is moved into AbstractReplicatedLog as a package-private method, with
more performant implementation. Tests now use another version of
clear().

removeFromAndPersist() is renamed to trimToReceive() and documented to
be called from Follower only. It is also marked for future evolution,
as it should contain essentially all of Follower.processAppendEntries().

Since this is the second construct we will be removing with Pekko
persiistence, add a common marker interface: PekkoPersistenceContract
implemented by both DeleteEntries and UpdateElectionTerm.

JIRA: CONTROLLER-2137
Change-Id: Iebe42d7f8ae82714b390ecbd0b51d75a78cc54a0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>