Split out EntryStore.markLastApplied()
Rather than going to through an async message, mark the last applied
index through persistence.

This eliminate one of the possibilities for persistAsync() and also
reduces latency and improves accuracy.

Since we are no longer sending ApplyJournalEntries to self, we also need
to switch tests to synchronize on applyIndex instead.

JIRA: CONTROLLER-2137
Change-Id: Id5aa178aa721b5086163d7b721b9447b80724d7e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>