Allow Snapshot state to be null
We are special-casing EmptyState to act as a placeholder for when we do
not want to propagate initial state.

This will not work SnapshotStore, as we cannot smuggle EmptyState
handling into deserialization handling.

Allow null Snapshot.State, mapping it to an empty stream -- and require
non-null States to emit at least one byte.

JIRA: CONTROLLER-2134
Change-Id: I13b6d702b61b4c8d1e7d77299ba872669c478d0a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>