Add EntryWriter.checkpoint()
We are going to be using hybrid data storage with RaftJournal, for which
we'll need a bit more control over which parts can and cannot be
discarded.

This patch adds EntryWriter.checkpoint(), which just moves to the next
segment -- thus allowing users to get better control over what data can
be purged.

JIRA: CONTROLLER-2137
Change-Id: I32256c44c9ff078721aca5263781fc93e6351343
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>