WIP: Switch EntryStore to use RaftJournal directly
Use a RaftJournal directly to store entries without the need to go
through Pekko Persistance and rely on it to do the Right Thing(tm).

We refactor EntryStore and ReplicateLogImpl to align with EntryJournal
lifecycle and overall way of doing thigs. A key part of that is taking a
snapshot once Pekko Persistence completes, which means some tests need
to be updated to take into account this snapshot being created, even
when persistence is off.

[ERROR] Failures:
[ERROR]   ReplicationAndSnapshotsWithLaggingFollowerIntegrationTest.testLeaderSnapshotTriggeredByMemoryThresholdExceededWithLaggingFollower:516 expected null, but was:<SnapshotFileV1{lastIncluded=EntryInfo[index=-1, term=-1], timestamp=2025-07-06T12:25:59.335761868Z, file=/tmp/org.opendaylight.controller.cluster.raft.ReplicationAndSnapshotsWithLaggingFollowerIntegrationTest11265296667355871101/leader373/snapshot-00000000686A6B57-140351CC.v1, size=221, entryCompress=NONE, stateCompress=NONE}>
[ERROR]   ClusterAdminRpcServiceTest.testRemoveShardReplica:458 Peers for shard cars expected:<[member-1-shard-cars-config_testRemoveShardReplica]> but was:<[member-3-shard-cars-config_testRemoveShardReplica, member-1-shard-cars-config_testRemoveShardReplica]>

JIRA: CONTROLLER-2137
Change-Id: Iaeac853e19966a2fde17450802682967a5e7e01c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>