Rework objectToBytes() overflow signalling
We are about to switch serialization in a way where it is much more
likely we will signal a buffer overflow, i.e. it will be normal if there
are fewer than maxEntrySize bytes left in a segment.

While the exact tuning remains to be determined and the exact frequency
will depend on the distribution of sizes in the workload, we can expect
the frequency to be higher than 1.0E-3, i.e. more than once in a thousand
invocations, just because the serialized entry size happens to be near
the maxEntrySize.

Such a frequency borders on not being exceptional enough and we really
are using EOFException for flow control, with the wringle of
EOFException being an IOException, which signals an irrecoverable error.

For detailed explanation see https://shipilev.net/blog/2014/exceptional-performance/, with the added
wrinkle that EOFException is a subclass of IOException, i.e. we

This patch changes objectToBytes() contract to return an explicit
boolean to differentiate success/overflow cases, eliminating the need to
catch EOFException and opting for an explicit check. The expectation is
that most calls will end up succeeding and the overflows will just take
an out-of-line slow path without relying on exceptions and their class
hierarchy checks.

JIRA: CONTROLLER-2137
Change-Id: I2f63671e932ccbd7aefeae4531219c1eba355f3a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>