Inline incremental recovery snapshots
Time-based increment snapshots should not be going through
SnapshotManager, as that performs things asynchronously, whereas we
really need the write the snapshot synchronously with the current
ReplicatedLog state, without it moving forward due to Pekko recovery
progressing.

This paves the way for EntryStore migration, removing the need for
SnapshotManager to talk to EntryStore -- which it currently does to
clean up Pekko.

JIRA: CONTROLLER-2137
Change-Id: I2cd770e5773de886a1a2502b33b25022ee2ab5ec
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>