AbstractReplicatedLog should be generic
ReplicatedLog implementations need additional internal structures to
keep track of entries: note how we use SimpleReplicatedLogEntry to track
persistence.

Pull the type of stored entries into a generic argument, introducing
adoptEntry() to deal with the differences.

This also allows us to be less reliant on ReplicatedLogEntry and use
more of LogEntry.

JIRA: CONTROLLER-2137
Change-Id: I2869522184f470fa2d160e40d9b9008f2c94f65d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>