Add EntryJournalV1
We need to provide an alternative to Pekko persistence when it comes to
EntryStore. This is a rather tall ask, given that the capabilities
afforded to by RaftJournal lend themselves to a much more streamlined
lifecycle.

This patch splits out the raw interface to maintain LogEntry objects in
a RaftJournal, along with the house keeping.

JIRA: CONTROLLER-2137
Change-Id: I4d7d3364a5422a8933420200265477051247a3e7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>