Repackage scala3-library
Upstream publishes two automatic modules:
- scala.library in scala-library
- org.scala.lang.scala3.library in scala3-library_3

Unfortunately, they contain split packages, which ends up being pain in
both OSGi and JPMS. In order to solved that, we introduce our own
packaging, which combines the two artifacts.

Another wrinkle is that we have to suport both [2,3) and [3,4) imports
of shared packages. This is completely doable with bnd, but ends up
being a tad long-winded.

JIRA: CONTROLLER-2167
Change-Id: Icf40bfe3390bd6e332e29e7a266919e06cef0098
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>