WIP: Repackage scala-library
Upstream publishes two automatic modules:
- scala.library in scala-library
- org.scala.lang.scala3.library in scala3-library_3

Unfortunately, they contain split packages, which ends up being pain in
both OSGi and JPMS.

In OSGi, this problem would be solvable, if the latter were a fragment
-- that way it would just be combined with scala-library and everything
would be fine.

In JPMS, such an automated thing is not possible and one has to resort
to patch-module, as shown here:
https://users.scala-lang.org/t/gradle-7-6-jdk-19-and-scala-3-java-lang-classnotfoundexception-scala-predef-even-though-scala-library-2-13-10-jar-is-in-classpath/8972

The Scala community will need to eventually fix this as part of their
Java interop story, but in the meantime we monkey-patch the problem and
create a combined bundle, just as we do for Pekko.

JIRA: CONTROLLER-2167

Change-Id: Icf40bfe3390bd6e332e29e7a266919e06cef0098
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>