Fix OSGiDatastoreContextIntrospectorFactory lifecycle
The injected serializer should be dynamically updated, so that the
entire datastore does not flap no model context updates.

This is fine as we are using it to just perform a one-time
serialization operation -- so do not particularly care about the speed,
so are using the easier AtomicReference indirection.

This also makes the two user-facing classes non-@Beta.

JIRA: CONTROLLER-2172
Change-Id: Ic0f362f3c691c778a48d56cee58c4a569f49da4f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>