Use service injection to drive RaftPolicy
This patch moves RaftPolicy to raft-spi, making it more widely
accessible. We then introduce WellKnownRaftPolicy, which centralizes the
three production policies we are shipping.

We then introduce a RaftPolicyResolver, which can use symbolic names to
look up the correct policy based on the instances available -- either
through OSGi SR or via ServiceLoader.

This eliminates the need to directly instantiate classes via
Class.forName() and, by extension, the need to use
DynamicImport-Package.

JIRA: CONTROLLER-2149
Change-Id: I5c42939a88136bc9313a7680472873b4ccbc1ddd
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>