De-anonymize ReadyCohort callbacks
We have 6 anonymous classes handling dispatch, which makes debugging a
tad difficult. This adds explicitly-named callbacks, so they can be
readily identified.

This also drops the number of actual classes to 5, as the doCommit()
part ends up being handled by same class.

This also results in "PreCommit" being spelled correctly in the failure
path, as well as "Commit" and "DoCommit" using the same "DoCommit"
message.

We keep the abort() case as is for now.

JIRA: CONTROLLER-2131
Change-Id: I4c081fd2d7b57cce68f06ebb38eaf064ac9fab1e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>